SCAL-94877 - Added autoLogin flag (#330)
* SCAL-94877 - Added autoLogin flag

* Set default autoLogin to true

* embed.ts -> ts-embed.ts

* Fixed lint issues

* Decreased coverage to commit changes
